+++ Source/Runtime/Core/Public/Math.h
@@ -76,7 +76,7 @@ struct Vector3
     Vector3 operator-(const Vector3& rhs) const { return {x - rhs.x, y - rhs.y, z - rhs.z}; }
     Vector3 operator*(float s) const { return {x * s, y * s, z * s}; }
 
-    float Length() const { return std::sqrt(x * x + y * y + z * z); }
+    float Length() const { return sqrtf(x * x + y * y + z * z); }
     float LengthSq() const { return x * x + y * y + z * z; }
 
     Vector3 Normalized() const
